add construct_message_to_log
